+++ libs/cardano-ledger-core/src/Cardano/Ledger/Language.hs
@@ -1,8 +1,10 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE KindSignatures #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE OverloadedStrings #-}
@@ -14,19 +16,42 @@
 -- | This module provides data structures and operations for talking about
 --     Non-native Script languages. It is expected that new languages (or new
 --     versions of old languages) will be added here.
-module Cardano.Ledger.Language where
+module Cardano.Ledger.Language (
+  -- * Plutus Script
+  Plutus (..),
+  BinaryPlutus (..),
+
+  -- * Value level Plutus Language version
+  Language (..),
+  mkLanguageEnum,
+  languageToText,
+  languageFromText,
+  nonNativeLanguages,
+  guardPlutus,
+
+  -- * Type level Plutus Language version
+  SLanguage (..),
+  IsLanguage (..),
+  fromSLanguage,
+  toSLanguage,
+  withSLanguage,
+) where
 
 import Cardano.Ledger.Binary (
+  Annotator,
   DecCBOR (..),
   Decoder,
   EncCBOR (..),
+  FromCBOR,
+  ToCBOR,
   decodeEnumBounded,
   encodeEnum,
   natVersion,
   unlessDecoderVersionAtLeast,
  )
+import Cardano.Ledger.SafeHash (SafeToHash (..))
 import Cardano.Ledger.TreeDiff (ToExpr (..))
-import Control.DeepSeq (NFData (..))
+import Control.DeepSeq (NFData (..), rwhnf)
 import Data.Aeson (
   FromJSON (parseJSON),
   FromJSONKey (fromJSONKey),
@@ -37,11 +62,39 @@ import Data.Aeson (
   withText,
  )
 import Data.Aeson.Types (toJSONKeyText)
+import qualified Data.ByteString.Base64 as B64 (encode)
+import Data.ByteString.Short (ShortByteString, fromShort)
 import Data.Ix (Ix)
 import Data.Text (Text)
 import Data.Typeable (Typeable)
 import GHC.Generics (Generic)
 import NoThunks.Class (NoThunks)
+
+data Plutus = Plutus
+  { plutusLanguage :: !Language
+  , plutusScript :: !BinaryPlutus
+  }
+  deriving stock (Eq, Ord, Show, Generic)
+
+-- | Already in Normal Form
+instance NFData Plutus where
+  rnf = rwhnf
+
+instance NoThunks Plutus
+
+-- | Binary representation of a Plutus script.
+newtype BinaryPlutus = BinaryPlutus {unBinaryPlutus :: ShortByteString}
+  deriving stock (Eq, Ord, Generic)
+  deriving newtype (ToCBOR, FromCBOR, EncCBOR, DecCBOR, NFData, NoThunks)
+
+instance Show BinaryPlutus where
+  show = show . B64.encode . fromShort . unBinaryPlutus
+
+instance DecCBOR (Annotator BinaryPlutus) where
+  decCBOR = pure <$> decCBOR
+
+instance SafeToHash BinaryPlutus where
+  originalBytes (BinaryPlutus binaryBlutus) = fromShort binaryBlutus
 
 -- | Non-Native Script language. This is an Enumerated type.
 -- This is expected to be an open type. We will add new Constuctors
@@ -155,6 +208,13 @@ toSLanguage lang
   where
     thisLanguage :: SLanguage l
     thisLanguage = isLanguage
+
+withSLanguage :: Language -> (forall l. IsLanguage l => SLanguage l -> a) -> a
+withSLanguage l f =
+  case l of
+    PlutusV1 -> f SPlutusV1
+    PlutusV2 -> f SPlutusV2
+    PlutusV3 -> f SPlutusV3
 
 -- | Prevent decoding a version of Plutus until
 -- the appropriate protocol version.