+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Version history for `cardano-ledger-core`
 
+## 1.5.0.0
+
+* Added `withSLanguage`
+* Move `BinaryPlutus` from `cardano-ledger-alonzo` and changed its `Show` instance to
+  display as base64 encoding.
+* Added `Plutus`
+* Changed `Phase2Script` constructor of the `PhaseScript` type. It now accepts new
+  `Plutus` type instead of `Language` and `ShortByteString`
+
 ## 1.4.0.0
 
 * Added `DRep`, `DRepCredential`